[Fire Mage] Misc Updates
Misc updates for Fire Mage support.
- Combined Combustion modules
- Replaced some references to shadowlands functions
- Removed some references to Infernal Cascade
- Replaces some buff IDs for various shadowlands specific things that returned as talents
- Added Shifting Power Ticks as CASTS THAT ARENT CASTS

There is still more work to do on Fire, but this is a good start and helps with some of the more immediate issues.
