+++ src/analysis/retail/mage/frost/core/WintersChill.tsx
@@ -145,13 +145,6 @@ class WintersChill extends Analyzer {
       return;
     }
 
-    //If the player is Venthyr and the target has Mirrors of Torment, then we need to ignore the precast
-    //We are doing this here because it is the only events where we can get the target debuffs
-    if (enemy.hasBuff(SPELLS.MIRRORS_OF_TORMENT.id)) {
-      this.preCastIgnored = true;
-      debug && this.log('PRE CAST IGNORED');
-    }
-
     this.shatteredCasts += 1;
     this.shatteredCastIDs.push(event.ability.guid);
   }
