+++ src/analysis/retail/mage/shared/constants.tsx
@@ -48,7 +48,7 @@ export const ARCANE_POWER_ALLOWED_CASTS = [
 ];
 
 //Mage Generic
-export const SHIFTING_POWER_MS_REDUCTION_PER_TICK = 2500;
+export const SHIFTING_POWER_MS_REDUCTION_PER_TICK = 3000;
 export const SHIFTING_POWER_REDUCTION_SPELLS = [
   SPELLS.FROST_NOVA,
   SPELLS.BLINK,
