+++ src/common/SPELLS/shadowlands/covenants/mage.ts
@@ -7,31 +7,6 @@ const covenants = spellIndexableList({
     name: 'Radiant Spark',
     icon: 'ability_bastion_mage',
   },
-
-  //Venthyr
-  MIRRORS_OF_TORMENT: {
-    id: 314793,
-    name: 'Mirrors of Torment',
-    icon: 'ability_revendreth_mage',
-  },
-
-  //Necrolord
-  DEATHBORNE: {
-    id: 324220,
-    name: 'Deathborne',
-    icon: 'ability_maldraxxus_mage',
-  },
-
-  //Night Fae
-  SHIFTING_POWER: {
-    id: 314791,
-    name: 'Shifting Power',
-    icon: 'ability_ardenweald_mage',
-  },
-  SHIFTING_POWER_TICK: {
-    id: 325130,
-    name: 'Shifting Power',
-    icon: 'ability_ardenweald_mage',
-  },
+  //Removed all other Shadowlands Covenant Spells
 });
 export default covenants;
