+++ src/parser/shared/normalizers/Channeling.ts
@@ -47,7 +47,7 @@ class Channeling extends EventsNormalizer {
     buffChannelSpec(SPELLS.SOUL_INFUSION.id), // fight channel from Sun King's Salvation - see in this log: https://wowanalyzer.com/report/g4Pja6pLHnmQtbvk/32-Normal+Sun+King's+Salvation+-+Kill+(10:14)/Pjurbo/standard/events
     // Mage
     buffChannelSpec(TALENTS_MAGE.EVOCATION_TALENT.id),
-    buffChannelSpec(SPELLS.SHIFTING_POWER.id),
+    buffChannelSpec(TALENTS_MAGE.SHIFTING_POWER_TALENT.id),
     nextCastChannelSpec(TALENTS_MAGE.ARCANE_MISSILES_TALENT.id),
     // Priest
     buffChannelSpec(TALENTS_PRIEST.DIVINE_HYMN_TALENT.id),
